Move ExecutorSampler decoder
