+++ qiskit_ibm_runtime/decoders/executor_sampler/__init__.py
@@ -10,4 +10,4 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""V0.1 post processor for executor-based SamplerV2"""
+"""Result decoder for executor-based Sampler."""
