+++ qiskit_ibm_runtime/decoders/executor_sampler/post_processor_v0_1.py
@@ -18,7 +18,7 @@ from typing import TYPE_CHECKING, cast
 
 from qiskit.primitives import PrimitiveResult
 
-from ..converters import quantum_program_result_to_primitive_result
+from .converters import quantum_program_result_to_primitive_result
 from .utils import executor_metadata_to_sampler_metadata, flatten_twirling_axes
 
 if TYPE_CHECKING:
