+++ qiskit_ibm_runtime/decoders/noise_learner_v3/converters.py
@@ -12,18 +12,22 @@
 
 """Transport conversion functions."""
 
-from qiskit.quantum_info import QubitSparsePauliList
+from __future__ import annotations
 
-from ibm_quantum_schemas.noise_learner_v3.version_0_2 import (
-    NoiseLearnerV3ResultModel,
-    NoiseLearnerV3ResultsModel,
-    ParamsModel,
-)
+from typing import TYPE_CHECKING
+
+from qiskit.quantum_info import QubitSparsePauliList
 
 from ...noise_learner_v3.noise_learner_v3_result import (
     NoiseLearnerV3Result,
     NoiseLearnerV3Results,
 )
+
+if TYPE_CHECKING:
+    from ibm_quantum_schemas.noise_learner_v3.version_0_2 import (
+        NoiseLearnerV3ResultsModel,
+    )
+
 
 def noise_learner_v3_result_from_0_1(
     model: NoiseLearnerV3ResultsModel,
