+++ qiskit_ibm_runtime/decoders/quantum_program/decoder.py
@@ -89,7 +89,7 @@ class QuantumProgramResultDecoder(ResultDecoder):
 
         if semantic_role == "sampler_v2":
             # TODO: Circular import issue. Consider changing file structure.
-            from ...executor_sampler.post_processors.post_processor_v0_1 import (
+            from ..executor_sampler.post_processor_v0_1 import (
                 sampler_v2_post_processor_v0_1,
             )
