+++ qiskit_ibm_runtime/noise_learner_v3/__init__.py
@@ -33,7 +33,7 @@ Classes
 """
 
 from .noise_learner_v3 import NoiseLearnerV3
-from .noise_learner_v3_result import (  # type: ignore[attr-defined]
+from .noise_learner_v3_result import (
     NoiseLearnerV3Result,
     NoiseLearnerV3Results,
 )
